Common UploadPolicyDropzone
